requirements.txt
Flask
Flask-SQLAlchemy
numpy
scikit-learn
xgboost
lightgbm
joblib
pandas
psycopg2-binary